+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivity.kt
@@ -760,6 +760,7 @@ class MainActivity :
             upNextQueue to useRssArtwork
         }
             .onEach { (upNextQueue, useRssArtwork) ->
+                updatePlaybackStateDeselectedChapterIndices(upNextQueue)
                 binding.playerBottomSheet.setUpNext(
                     upNext = upNextQueue,
                     theme = theme,
@@ -885,6 +886,21 @@ class MainActivity :
                 }
             }
         })
+    }
+
+    private fun updatePlaybackStateDeselectedChapterIndices(upNextQueue: UpNextQueue.State) {
+        (upNextQueue as? UpNextQueue.State.Loaded)?.let {
+            val lastPlaybackState = viewModel.lastPlaybackState
+            val currentEpisodeDeselectedChapterIndicesChanged = playbackManager.getCurrentEpisode()?.let { currentEpisode ->
+                currentEpisode.uuid == it.episode.uuid &&
+                    lastPlaybackState != null &&
+                    it.episode.deselectedChapters !=
+                    lastPlaybackState.chapters.getList().filterNot { it.selected }.map { it.index }
+            } ?: false
+            if (currentEpisodeDeselectedChapterIndicesChanged) {
+                playbackManager.updatePlaybackStateDeselectedChapterIndices()
+            }
+        }
     }
 
     override fun whatsNewDismissed(fromConfirmAction: Boolean) {