+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/converter/ChapterIndicesConverter.kt
@@ -5,11 +5,8 @@ import au.com.shiftyjelly.pocketcasts.models.entity.ChapterIndices
 
 class ChapterIndicesConverter {
     @TypeConverter
-    fun fromString(value: String?): ChapterIndices {
-        val list = value?.split(",")?.mapNotNull { it.toIntOrNull() } ?: emptyList()
-        return ChapterIndices(list)
-    }
+    fun fromString(value: String?) = ChapterIndices.fromString(value)
 
     @TypeConverter
-    fun toString(indices: ChapterIndices) = indices.joinToString(",")
+    fun toString(indices: ChapterIndices) = ChapterIndices.toString(indices)
 }